modified model to new api

+++ lib/presentation/screens/food_detail_page/food_detail_page.dart
@@ -146,7 +146,7 @@ class FoodDetailPage extends StatelessWidget {
                   SizedBox(
                     width: 10,
                   ),
-                  Text("Rs: " + food.price.toInt().toString(),
+                  Text("Rs: " + food.price.toString(),
                       style: TextStyle(
                           fontSize: 20,
                           color: Colors.white,
